Refactor to SessionService
