+++ src/Secure/WebservicesAuthentication.php
@@ -5,6 +5,7 @@ namespace PhpTwinfield\Secure;
 use PhpTwinfield\Enums\Services;
 use PhpTwinfield\Exception;
 use PhpTwinfield\Services\BaseService;
+use PhpTwinfield\Services\SessionService;
 
 class WebservicesAuthentication extends AuthenticatedConnection
 {
@@ -56,57 +57,9 @@ class WebservicesAuthentication extends AuthenticatedConnection
             return;
         }
 
-        $loginService = new class extends BaseService {
+        $sessionService = new SessionService();
 
-            private const LOGIN_OK = "Ok";
-
-            /**
-             * @param string $username
-             * @param string $password
-             * @param string $organization
-             * @return string[]
-             * @throws Exception
-             */
-            public function getSessionIdAndCluster(string $username, string $password, string $organization): array
-            {
-                $response = $this->Logon([
-                    "user"         => $username,
-                    "password"     => $password,
-                    "organisation" => $organization,
-                ]);
-
-                $result = $response->LogonResult;
-
-                // Check response is successful
-                if ($result !== self::LOGIN_OK) {
-                    throw new Exception("Failed logging in using the credentials, result was \"{$result}\".");
-                }
-
-                // Response from the logon request
-                $loginResponse = $this->__getLastResponse();
-
-                // Make a new DOM and load the response XML
-                $envelope = new \DOMDocument();
-                $envelope->loadXML($loginResponse);
-
-                // Gets SessionID
-                $sessionIdElements = $envelope->getElementsByTagName('SessionID');
-                $sessionId = $sessionIdElements->item(0)->textContent;
-
-                // Gets Cluster URL
-                $clusterElements = $envelope->getElementsByTagName('cluster');
-                $cluster = $clusterElements->item(0)->textContent;
-
-                return [$sessionId, $cluster];
-            }
-
-            final protected function WSDL(): string
-            {
-                return "https://login.twinfield.com/webservices/session.asmx?wsdl";
-            }
-        };
-
-        [$this->sessionID, $this->cluster] = $loginService->getSessionIdAndCluster($this->username, $this->password, $this->organization);
+        [$this->sessionID, $this->cluster] = $sessionService->getSessionIdAndCluster($this->username, $this->password, $this->organization);
     }
 
     protected function getSoapHeaders()
